Add collapsible side panels and split oversized UI modules.
Group Studio Monitor and Test Patterns side-panel stats so they can collapse, make the dividers resizable, and split the 2000-line UI files for maintainability.
